backend/requirements.txt
Django==3.0.8

djangorestframework==3.11.0
django-cors-headers==3.4.0

celery==4.4.2
redis==3.5.3
django-redis==4.12.1

gunicorn==20.0.4

psycopg2==2.8.5
psycopg2-binary==2.8.5

mixer==6.1.3
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
snapshottest==0.5.1
